Re-support DISTINCT-qualified Window Aggregate in Orca (#17113)
This was enabled in 5X, but was lost in the 6X merge and Orca fell back to Planner. This was re-enabled in the executor in 6523b43, but was not added for Orca. 386d542 enables the fallback in Orca.

In 5X, Orca and planner would generate the following plan for the query
create table t2 (a int, b int, c int) distributed by (c);
explain select a, count( distinct b) over (partition by a) from t2;

```
 Gather Motion 3:1  (slice2; segments: 3)  (cost=0.00..1293.01 rows=20 width=12)
   ->  Sequence  (cost=0.00..1293.01 rows=7 width=12)
         ->  Shared Scan (share slice:id 2:0)  (cost=0.00..431.00 rows=7 width=1)
               ->  Materialize  (cost=0.00..431.00 rows=7 width=1)
                     ->  Redistribute Motion 3:3  (slice1; segments: 3)  (cost=0.00..431.00 rows=7 width=42)
                           Hash Key: t2.a
                           ->  Table Scan on t2  (cost=0.00..431.00 rows=7 width=42)
         ->  Hash Join  (cost=0.00..862.00 rows=7 width=12)
               Hash Cond: NOT share0_ref3.a IS DISTINCT FROM share0_ref2.a
               ->  GroupAggregate  (cost=0.00..431.00 rows=4 width=12)
                     Group By: share0_ref3.a
                     ->  Sort  (cost=0.00..431.00 rows=7 width=8)
                           Sort Key: share0_ref3.a
                           ->  Shared Scan (share slice:id 2:0)  (cost=0.00..431.00 rows=7 width=8)
               ->  Hash  (cost=431.00..431.00 rows=7 width=4)
                     ->  Shared Scan (share slice:id 2:0)  (cost=0.00..431.00 rows=7 width=4)
 Optimizer status: PQO version 3.117.0
(17 rows)
```

Now, we'll generate the following plan in both Orca and Planner:
```
test=# explain select a, count( distinct b) over (partition by a) from t2;
                                              QUERY PLAN
------------------------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..431.00 rows=1 width=12)
   ->  WindowAgg  (cost=0.00..431.00 rows=1 width=12)
         Partition By: a
         ->  Sort  (cost=0.00..431.00 rows=1 width=8)
               Sort Key: a
               ->  Redistribute Motion 3:3  (slice2; segments: 3)  (cost=0.00..431.00 rows=1 width=8)
                     Hash Key: a
                     ->  Seq Scan on t2  (cost=0.00..431.00 rows=1 width=8)
 Optimizer: GPORCA
(9 rows)
```
The windistinct flag is passed through the WindowAgg node, and the distinct operation is handled within the executor.
This was already done in the Orca code, and the logic was already implemented. However, we still had the preprocessing logic. We simply needed to stop doing this extra logic within the optimizer (in the preprocessor) since it's now done in the executor.

Basic performance tests show that this implementation is also faster than the 5X plan for both Orca and Planner when there is a PARTITION BY clause for the window function